Created a testing module to clean things up
